requirements.txt
Flask==1.1.1
PyMySQL==0.9.3
pandas==0.25.3
pytz==2019.3
python-dateutil==2.8.1
Scrapy==1.8.0